pinakdrf/ibps_api/requirement.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3